tbc: move RPC tests to a separate test file
